Enabled network connections from anywhere.
By default the pom.xml file is configured to only allow network connections
from localhost. Changing the source to 0.0.0.0, enables the web server to
receive network connections from any IP address. This is an ok setting for a
fully public web site, but it must be tought about on every case.